Add librespeed to gateways
Add librespeed frontend and backend containers to the gateways. Also add
a bunch of DNS entries, so that the backends can be selected from the
frontend by IP version. Caddy is used as reverse proxy to get a correct
X-Forwarded-For header, since user space networking in podman does not
see the external IP.